+++ src/sn/EventListenerCtn.ts
@@ -5,9 +5,9 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-import {BaseTexture, utils} from 'pixi.js';
+import {type Container, utils} from 'pixi.js';
 
-type IEmitter = BaseTexture | utils.EventEmitter | {
+type IEmitter = utils.EventEmitter | {
 	addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
 	removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
 }
@@ -15,19 +15,11 @@ type IEmitter = BaseTexture | utils.EventEmitter | {
 export class EventListenerCtn {	// リソースリーク対策
 	#aOffEvt	: {(): void}[]	= [];
 
+	addC(ed: Container, type: string, fnc: (e: any)=> void, ctx: any = {}): void {
+		ed.on(type, fnc, ctx);
+		this.#aOffEvt.push(()=> ed.off(type, fnc, ctx));
+	}
 	add(ed: IEmitter, type: string, fnc: (e: any)=> void, ctx: any = {}): void {
-		if (ed instanceof BaseTexture) {
-			switch (type) {
-			case 'loaded':
-			case 'update':
-			case 'error':
-			case 'dispose':
-				ed.on(type, fnc, ctx);
-				this.#aOffEvt.push(()=> ed.off(type, fnc, ctx));
-				break;
-			}
-			return;
-		}
 		if (ed instanceof utils.EventEmitter) {
 			ed.on(type, fnc, ctx);
 			this.#aOffEvt.push(()=> ed.off(type, fnc, ctx));